Added reverse DNS lookup to ssh-add-known-host script
If the user provides a hostname the script will now
use `host` to do a reverse DNS lookup and get the IP
address. When the SSH public key is added to known_hosts
both the hostname and IP address are included when
available.

Also, the ssh-keyscan -H option is no longer used to hash
the public key, as it made removing existing entries in
the known_hosts file unpredictable.