T408: clean up the AMI build scripts and retarget them to 1.2.0.
In particular,
Use Debian Jessie instead of ancient Ubuntu Precise.

Add a step to install parted, gnupg, and e2fsprogs if they are not installed:
it's not guaranteed that the Debian AMI has those.
Do not create a separate VPC for the build instance, use the first available subnet
in the default VPC.

Replace shell commands with ansible module calls whenever possible.
Remove dependencies on deprecated ansible modules.

Use config file template instead of in-place editing of the default config from the image.

Remove unused files.